+++ bean-searcher/src/main/java/com/ejlchina/searcher/implement/DefaultMetaResolver.java
@@ -102,8 +102,7 @@ public class DefaultMetaResolver implements MetaResolver {
             beanMeta.addFieldMeta(wrapper.field.getName(), fieldMeta);
         }
         if (beanMeta.getFieldCount() == 0) {
-            throw new SearchException("[" + beanClass.getName() + "] is not a valid SearchBean, because there is no field mapping to database.\n" +
-                    "Please refer https://bs.zhxu.cn/guide/latest/bean.html#%E7%9C%81%E7%95%A5-dbfield for help.");
+            throw new SearchException("[" + beanClass.getName() + "] is not a valid SearchBean, because there is no field mapping to database. Please see https://bs.zhxu.cn/guide/latest/bean.html#%E7%9C%81%E7%95%A5-dbfield for help.");
         }
         return beanMeta;
     }