+++ CHANGELOG.md
@@ -11,11 +11,13 @@
   * 优化：参数构建器新增 `asc(boolean sure)` 与 `desc(boolean sure)` 方法
   * 优化：参数构建器新增 `putAll(Map<String, ?> params)` 方法
   * 优化：当分页尺寸小于等于 `0` 时，不执行列表查询
+  * 优化：标准化异常提示信息，全部英文化
 * Bean Searcher Boot Starter
   * 新增 `bean-searcher.params.filter.max-para-map-size` 配置项，默认 `150`
   * 新增 `bean-searcher.params.group.max-expr-length` 配置项，默认 `50`
   * 新增 `bean-searcher.params.pagination.max-allowed-offset` 配置项，默认 `20000`
   * 新增配置项校验：`bean-searcher.params.pagination.default-size` 的值不能比 `bean-searcher.params.pagination.max-allowed-size` 大，且都必须大于 `0`
+  * 优化：标准化异常提示信息，全部英文化
 
 # v3.8.0 @ 2022-07-23